@@ -20,15 +20,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "4.2.4",
-    "@angular/common": "4.2.4",
-    "@angular/compiler": "4.2.4",
-    "@angular/core": "4.2.4",
-    "@angular/forms": "4.2.4",
-    "@angular/http": "4.2.4",
-    "@angular/platform-browser": "4.2.4",
-    "@angular/platform-browser-dynamic": "4.2.4",
-    "@angular/router": "4.2.4",
+    "@angular/animations": "5.2.0",
+    "@angular/common": "5.2.0",
+    "@angular/compiler": "5.2.0",
+    "@angular/core": "5.2.0",
+    "@angular/forms": "5.2.0",
+    "@angular/http": "5.2.0",
+    "@angular/platform-browser": "5.2.0",
+    "@angular/platform-browser-dynamic": "5.2.0",
+    "@angular/router": "5.2.0",
     "core-js": "2.4.1",
     "font-awesome": "4.7.0",
     "foundation-sites": "6.4.3",
@@ -36,31 +36,30 @@
     "lodash.random": "3.2.0",
     "ngx-pagination": "3.0.3",
     "ngx-pipes": "2.0.5",
-    "rxjs": "5.4.2",
-    "zone.js": "0.8.14"
+    "rxjs": "5.5.6",
+    "zone.js": "0.8.19"
   },
   "devDependencies": {
-    "@angular/cli": "1.3.1",
-    "@angular/compiler-cli": "4.2.4",
-    "@angular/language-service": "4.2.4",
-    "@types/jasmine": "2.5.53",
+    "@angular/cli": "1.7.4",
+    "@angular/compiler-cli": "5.2.0",
+    "@angular/language-service": "5.2.0",
+    "@types/jasmine": "2.8.3",
     "@types/jasminewd2": "2.0.2",
     "@types/lodash": "4.14.74",
     "@types/node": "6.0.60",
-    "codelyzer": "3.1.1",
+    "codelyzer": "4.0.1",
     "gh-pages": "1.1.0",
     "http-server": "0.10.0",
-    "jasmine-core": "2.6.2",
-    "jasmine-spec-reporter": "4.1.0",
-    "karma": "1.7.0",
-    "karma-chrome-launcher": "2.1.1",
-    "karma-cli": "1.0.1",
+    "jasmine-core": "2.8.0",
+    "jasmine-spec-reporter": "4.2.1",
+    "karma": "2.0.0",
+    "karma-chrome-launcher": "2.2.0",
     "karma-coverage-istanbul-reporter": "1.2.1",
     "karma-jasmine": "1.1.0",
     "karma-jasmine-html-reporter": "0.2.2",
     "protractor": "5.1.2",
-    "ts-node": "3.2.0",
-    "tslint": "5.3.2",
-    "typescript": "2.3.3"
+    "ts-node": "4.1.0",
+    "tslint": "5.9.1",
+    "typescript": "2.5.3"
   }
 }
