@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Http } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 import 'rxjs/add/operator/toPromise';
 
 import { IPost } from 'app/shared/models/post.interface';
@@ -10,7 +10,7 @@ import { ConstantsApiService } from 'app/shared/constants-api.service';
 export class PostsService {
 
     constructor(
-        private http: Http,
+        private http: HttpClient,
         private errorService: ErrorService,
         private constantsApiService: ConstantsApiService
     ) { }
@@ -18,14 +18,14 @@ export class PostsService {
     getPosts(): Promise<IPost[]> {
         return this.http.get(this.constantsApiService.getJsonPlaceHolderUrl())
             .toPromise()
-            .then(data => data.json())  // this is necessary in this version of Http (but wont be in next version)
+            .then(data => data)
             .catch(error => this.errorService.rejectPromise('posts.service', error, true));
     }
 
     getPost(id: number): Promise<IPost> {
         return this.http.get(this.constantsApiService.getJsonPlaceHolderUrl() + id)
             .toPromise()
-            .then(data => data.json())  // this is necessary in this version of Http (but wont be in next version)
+            .then(data => data)
             .catch(error => this.errorService.rejectPromise('posts.service', error, true));
     }
 
